requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
bcrypt<4.0
